# platform/api/openapi.json
{
  "components": {
    "schemas": {
...
      "AtualizarFichaComando": {
        "additionalProperties": false,
        "properties": {
          "ator_id": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Ator Id",
            "type": "string"
          },
          "ficha": {
            "$ref": "#/components/schemas/FichaContrato"
          },
          "id": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Id",
            "type": "string"
          },
          "mesa_id": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Mesa Id",
            "type": "string"
          },
          "personagem_id": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Personagem Id",
            "type": "string"
          },
          "tipo": {
            "const": "atualizar_ficha",
            "default": "atualizar_ficha",
            "title": "Tipo",
            "type": "string"
          },
          "versao_esperada": {
            "minimum": 0.0,
            "title": "Versao Esperada",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "mesa_id",
          "ator_id",
          "versao_esperada",
          "personagem_id",
          "ficha"
        ],
        "title": "AtualizarFichaComando",
        "type": "object"
      },
...
      "FichaContrato": {
        "additionalProperties": true,
        "description": "Representação compatível de uma ficha transportada pela API.",
        "properties": {
          "armaduras": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Armaduras",
            "type": "array"
          },
          "armas": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Armas",
            "type": "array"
          },
          "atributos": {
            "additionalProperties": true,
            "title": "Atributos",
            "type": "object"
          },
          "efeitos_externos": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Efeitos Externos",
            "type": "array"
          },
          "outros": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Outros",
            "type": "array"
          },
          "pericias": {
            "additionalProperties": true,
            "title": "Pericias",
            "type": "object"
          },
          "personagem": {
            "additionalProperties": true,
            "title": "Personagem",
            "type": "object"
          },
          "personalidade": {
            "additionalProperties": true,
            "title": "Personalidade",
            "type": "object"
          }
        },
        "title": "FichaContrato",
        "type": "object"
      },
      "FichaSnapshot": {
        "properties": {
          "avisos": {
            "description": "Valores fora das regras ou do catálogo, mantidos até o Narrador corrigir.",
            "items": {
              "$ref": "#/components/schemas/ProblemaValidacao"
            },
            "title": "Avisos",
            "type": "array"
          },
          "ficha": {
            "$ref": "#/components/schemas/FichaContrato"
          },
          "mesa_id": {
            "title": "Mesa Id",
            "type": "string"
          },
          "personagem_id": {
            "title": "Personagem Id",
            "type": "string"
          },
          "tipo": {
            "default": "personagem",
            "description": "personagem, npc ou monstro: só personagens seguem limites e catálogo.",
            "title": "Tipo",
            "type": "string"
          },
          "versao": {
            "title": "Versao",
            "type": "integer"
          }
        },
        "required": [
          "mesa_id",
          "personagem_id",
          "versao",
          "ficha"
        ],
        "title": "FichaSnapshot",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PedidoAlteracaoResumo": {
        "properties": {
          "campos_alterados": {
            "items": {
              "type": "string"
            },
            "title": "Campos Alterados",
            "type": "array"
          },
          "estado": {
            "enum": [
              "pendente",
              "aprovado",
              "rejeitado"
            ],
            "title": "Estado",
            "type": "string"
          },
          "ficha_proposta": {
            "$ref": "#/components/schemas/FichaContrato"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "mesa_id": {
            "title": "Mesa Id",
            "type": "string"
          },
          "personagem_id": {
            "title": "Personagem Id",
            "type": "string"
          },
          "solicitante_id": {
            "title": "Solicitante Id",
            "type": "string"
          },
          "versao_base": {
            "title": "Versao Base",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "mesa_id",
          "personagem_id",
          "solicitante_id",
          "versao_base",
          "campos_alterados",
          "estado",
          "ficha_proposta"
        ],
        "title": "PedidoAlteracaoResumo",
        "type": "object"
      },
...
      "ProblemaValidacao": {
        "properties": {
          "campo": {
            "title": "Campo",
            "type": "string"
          },
          "mensagem": {
            "title": "Mensagem",
            "type": "string"
          }
        },
        "required": [
          "campo",
          "mensagem"
        ],
        "title": "ProblemaValidacao",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "API autoritativa da plataforma colaborativa Cursed.",
    "title": "Cursed — Plataforma Colaborativa",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/mesas/{mesa_id}/personagens/{personagem_id}": {
      "delete": {
        "operationId": "excluir_personagem_mesas__mesa_id__personagens__personagem_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "mesa_id",
            "required": true,
            "schema": {
              "title": "Mesa Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "personagem_id",
            "required": true,
            "schema": {
              "title": "Personagem Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "versao_esperada",
            "required": true,
            "schema": {
              "minimum": 0,
              "title": "Versao Esperada",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Excluir Personagem",
        "tags": [
          "Personagens e políticas"
        ]
      }
    },
...
    "/mesas/{mesa_id}/personagens/{personagem_id}/ficha": {
      "get": {
        "operationId": "ler_ficha_mesas__mesa_id__personagens__personagem_id__ficha_get",
        "parameters": [
          {
            "in": "path",
            "name": "mesa_id",
            "required": true,
            "schema": {
              "title": "Mesa Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "personagem_id",
            "required": true,
            "schema": {
              "title": "Personagem Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FichaSnapshot"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Ler Ficha",
        "tags": [
          "Fichas"
        ]
      },
      "put": {
        "operationId": "gravar_ficha_mesas__mesa_id__personagens__personagem_id__ficha_put",
        "parameters": [
          {
            "in": "path",
            "name": "mesa_id",
            "required": true,
            "schema": {
              "title": "Mesa Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "personagem_id",
            "required": true,
            "schema": {
              "title": "Personagem Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AtualizarFichaComando"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FichaSnapshot"
                }
              }
            },
            "description": "Successful Response"
          },
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PedidoAlteracaoResumo"
                }
              }
            },
            "description": "Accepted"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Gravar Ficha",
        "tags": [
          "Fichas"
        ]
      }
    },
...
  }
}